training/pyproject.toml
[tool.poetry]
name = "aapl"
version = "0.1.0"
description = ""
authors = ["Shuhei M. Yoshida <s_m_yoshida@nec.com>"]
readme = "README.md"
packages = [{ include = "aapl", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"

torch = [
  { version = "2.0.1+cu118", markers = "sys_platform == 'linux'", source = "pytorch-cu118" },
  { version = "2.0.1", markers = "sys_platform != 'linux'", source = "pypi" },
]
torchvision = [
  { version = "0.15.2+cu118", markers = "sys_platform == 'linux'", source = "pytorch-cu118" },
  { version = "0.15.2", markers = "sys_platform != 'linux'", source = "pypi" },
]
numpy = "1.24.3"
scipy = "1.10.1"

pandas = "^2.1.1"

joblib = "^1.3.2"
omegaconf = "^2.3.0"
pytorch-ignite = "^0.4.13"
scikit-learn = "^1.3.2"
tqdm = "^4.66.1"
mlflow = "^2.9.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.11.4"
flake8 = "^6.1.0"
mypy = "^1.5.1"
mdformat-gfm = "^0.3.5"
mdformat_frontmatter = "^0.4.1"
mdformat_footnote = "^0.1.1"
plotly = "^5.18.0"
nbformat = ">=4.2.0"
ptflops = "^0.7.2.2"


[tool.poetry.group.workaround.dependencies]
alembic = "1.12.1"
certifi = "2023.7.22"
colorlog = "6.7.0"
idna = "3.4"
mako = "1.2.4"
urllib3 = "2.0.7"

[[tool.poetry.source]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_decorated = false
disallow_any_generics = false
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
